perf: shrink Node from 72 to 64 bytes
Use uint32 for TreeKind, token.Kind and Node.len instead of uint/int,
reducing each from 8 to 4 bytes. Reorder fields so the three 4-byte
fields are adjacent, leaving only 4 bytes of alignment padding before
the next 8-byte-aligned field.

  // 72 bytes (before)          // 64 bytes (after)
  Kind      TreeKind/uint (8B)  Kind      TreeKind/uint32 (4B)
  TokenKind token.Kind/uint(8B) TokenKind token.Kind/uint32 (4B)
  Start     token.Position(16B) len       uint32 (4B)
  End       token.Position(16B) // 4 bytes padding
  Literal   string (16B)        Start     token.Position (16B)
  len       int (8B)            End       token.Position (16B)
                                Literal   string (16B)

Measured formatting the entire samples-graphviz directory (998 DOT
files, from the graphviz repository) with:

  time dotx fmt samples-graphviz/

Before:
  8.34 real, 14.77 user, 2.55 sys, 21131 ctx-switch
After:
  7.69 real, 13.43 user, 2.46 sys, 20748 ctx-switch

+++ parser.go
@@ -128,7 +128,7 @@ func (p *Parser) openNode(kind TreeKind) int {
 
 func (p *Parser) closeNode(i int) {
 	n := &p.tree.nodes[i]
-	n.len = len(p.tree.nodes) - i - 1
+	n.len = uint32(len(p.tree.nodes) - i - 1)
 	if n.len > 0 {
 		// Set Start/End from first and last descendant
 		first := p.tree.nodes[i+1]
